feat(console): update tournament page style

+++ packages/console/src/pages/tournament.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { computed, toRefs } from "vue"
+import { computed, toRefs, watchEffect } from "vue"
 import { useRoute } from "vue-router"
 import { useStore } from "../store/store"
 import http from "../utils/http"
@@ -8,6 +8,11 @@ import Markdown from "../components/common/markdown.vue"
 import UserLink from "../components/common/user-link.vue"
 import EventLink from "../components/common/event-link.vue"
 import { NotFoundError } from "../utils/not-found-error"
+// noinspection ES6UnusedImports
+import {
+  Account, AccountGroup, AccountMultiplePlus, AccountPlus, AccountRemove, CalendarEdit, Delete, ExitRun, FlagCheckered,
+  HumanGreeting, LeadPencil, LocationEnter, LocationExit, Pencil, Play,
+} from "mdue"
 
 const route = useRoute()
 const store = useStore()
@@ -26,8 +31,8 @@ let tournament = $ref(tournaments[0])
 
 document.title = `${ tournament.name } - ${ document.title }`
 
-if (tournament.status === Status.Finished)
-  tournament.teams.sort(({ rank: a }, { rank: b }) => a - b)
+// true if this is a solo tournament
+let solo = $(computed(() => tournament.teamSizeMax <= 1))
 
 // references the registration object of the current user
 let myself = $(computed(() => user.registrations.find(r => r.eventId === event.id)))
@@ -55,12 +60,35 @@ let teamsGroups = $(computed(() =>
 
 const isTeamFull = (team: Team) => team.members.length >= tournament.teamSizeMax
 let isTournamentFull = $(computed(() => tournament.teamCount >= tournament.teamCountMax))
+let enoughParticipants = $(computed(() => tournament.teamCount >= tournament.teamCountMin))
+
+// keep teams sorted
+watchEffect(() => {
+  // sort by rank when finished
+  if (tournament.status === Status.Finished)
+    tournament.teams.sort(({ rank: a }, { rank: b }) => a - b)
+
+  else {
+    if (!solo)
+      // show teams with the fewest members first
+      tournament.teams.sort(({ members: a }, { members: b }) => a.length - b.length)
+
+    // move the current user at the top
+    if (isRegistered) {
+      tournament.teams.sort(team => team === myTeam ? -1 : 1)
+      if (!solo)
+        myTeam?.members.sort(member => member.userId === user.id ? -1 : 1)
+    }
+  }
+
+  return tournament.teams
+})
 
 /**
  * Registers the current user to the tournament by creating a new team.
  */
 async function register() {
-  const name = tournament.teamSizeMax > 1 ? prompt("Enter team name:") : user.displayName
+  const name = solo ? user.displayName : prompt("Enter team name:")
   if (!name || !name.length) return
 
   const team = await http.post(`/tournaments/${ tournament.id }/teams`, { name }, Team)
@@ -82,7 +110,7 @@ async function unregister() {
  */
 async function createTeam() {
   let team: Team
-  if (tournament.teamSizeMax <= 1) {
+  if (solo) {
     const member = await promptMember()
     if (!member) return
     team = await http.post(`/tournaments/${ tournament.id }/teams`,
@@ -102,7 +130,8 @@ async function createTeam() {
  * @param team The target team
  * @param member The user to add to the team
  */
-async function addMember(team: Team, member: Registration) {
+async function addMember(team?: Team, member?: Registration) {
+  if (!team || !member) return
   await http.put(`/teams/${ team.id }/members/${ member.userId }`, {})
   // if this is the current user, we don't need to fetch their profile
   if (member.userId === user.id)
@@ -119,14 +148,17 @@ async function addMember(team: Team, member: Registration) {
  * @param team The target team
  * @param member The user to remove from the team
  */
-async function removeMember(team: Team, member: Registration) {
-  await http.delete(`/teams/${ team.id }/members/${ member.userId }`)
-  team.members = team.members.filter(({ userId }) => userId !== member.userId)
-  // also remove team if empty
-  if (!team.members.length)
-    tournament.teams.splice(tournament.teams.indexOf(team), 1)
+async function removeMember(team?: Team, member?: Registration) {
+  if (!team || !member) return
+  if (member === myself || confirm("Do you really want to remove " + member.user.displayName + "?")) {
+    await http.delete(`/teams/${ team.id }/members/${ member.userId }`)
+    team.members = team.members.filter(({ userId }) => userId !== member.userId)
+    // also remove team if empty
+    if (!team.members.length)
+      tournament.teams.splice(tournament.teams.indexOf(team), 1)
 
-  tournament.teamCount = fullTeams.length
+    tournament.teamCount = fullTeams.length
+  }
 }
 
 /**
@@ -146,9 +178,12 @@ async function renameTeam(team: Team) {
  * @param team The target team
  */
 async function deleteTeam(team: Team) {
-  await http.delete(`/teams/${ team.id }`)
-  tournament.teams.splice(tournament.teams.indexOf(team), 1)
-  tournament.teamCount = fullTeams.length
+  const name = solo ? team.members[0].user.displayName : team.name
+  if (confirm("Do you really want to remove " + name + "?")) {
+    await http.delete(`/teams/${ team.id }`)
+    tournament.teams.splice(tournament.teams.indexOf(team), 1)
+    tournament.teamCount = fullTeams.length
+  }
 }
 
 /**
@@ -181,6 +216,14 @@ async function addMemberPrompt(team: Team) {
 }
 
 /**
+ * Returns true if the current user can join a team.
+ * @param team The target team
+ */
+function canJoin(team: Team) {
+  return canRegister && !isRegistered && !isTeamFull(team)
+}
+
+/**
  * Locks registrations and start tournament.
  */
 async function startTournament() {
@@ -192,70 +235,360 @@ async function startTournament() {
 </script>
 
 <template lang="pug">
-h1 {{ tournament.name }}
-event-link(v-if="isOrganizer" to="edit-tournament" :params="{ name }") Edit
+.container(:class="{ [tournament.status]: true }")
+  header.tournament-header
+    .bg(:style="{ backgroundImage: `url(${tournament.background})` }")
+    .running(v-if="tournament.status === Status.Started")
+    .tournament-title
+      h1 {{ tournament.name }}
+      p.status(v-if="tournament.status === Status.Started") The tournament has started
+      p.status(v-if="tournament.status === Status.Finished") The tournament is over
 
-p.info {{ tournament.teamCount }}!{" "}
-  template(v-if="!tournamentStarted") / {{ tournament.teamCountMax }}
-  | !{" "}registered {{ tournament.teamSizeMax > 1 ? "teams" : "players" }}
+    .admin(v-if="isOrganizer")
+      span Administration
+      .actions
+        event-link.button(to="edit-tournament" :params="{ name }")
+          calendar-edit
+          span Edit
+        template(v-if="!tournamentStarted")
+          button.button(v-if="!isTournamentFull" @click="createTeam")
+            template(v-if="solo")
+              account-plus
+              span Add player
+            template(v-else)
+              account-multiple-plus
+              span New team
+          button.button.green(v-if="enoughParticipants" @click="startTournament")
+            play
+            span Start
+          button.button.green(v-else disabled title="Not enough participants")
+            play
+            span Start
+        event-link.button(v-else to="tournament-results" :params="{ name }")
+          template(v-if="tournament.status === Status.Started")
+            flag-checkered
+            span Finish
+          template(v-else)
+            lead-pencil
+            span Results
 
-template(v-if="!tournamentStarted")
-  template(v-if="canRegister")
-    button(v-if="isRegistered" @click="unregister") Unregister
-    button(v-if="!isRegistered && !isTournamentFull" @click="register") Register
+    .teamCount(:title="'Registered ' + (solo ? 'players' : 'teams')")
+      account(v-if="solo")
+      account-group(v-else)
+      .count
+        span.current(:class='{ low: !enoughParticipants }') {{ tournament.teamCount }}&nbsp;
+        span.max(v-if="!tournamentStarted") / {{ tournament.teamCountMax }}
 
-  template(v-if="isOrganizer")
-    button(v-if="!isTournamentFull" @click="createTeam")
-      | {{ tournament.teamSizeMax > 1 ? "Create team" : "Register player" }}
-    button(v-if="tournament.teamCount >= tournament.teamCountMin" @click="startTournament") Start
-    button(v-else disabled title="Not enough participants") Start
+    template(v-if="!tournamentStarted && canRegister")
+      button.button.unregister(v-if="isRegistered" @click="unregister")
+        exit-run
+        span Unregister
+      button.button.register(v-if="!isRegistered && !isTournamentFull" @click="register")
+        human-greeting
+        span Register
 
-event-link(v-else-if="isOrganizer" to="tournament-results" :params="{ name }")
-  | Enter results
+  .content
+    markdown.rules(:src="tournament.rules")
 
-markdown.rules(:src="tournament.rules")
+    .teamsGroup(v-for="(teams, index) in teamsGroups")
+      h2(v-if="tournament.status === Status.Finished") Results
+      h2(v-else-if="teamsGroups.length > 1 && index === 0") Incomplete teams
+      h2(v-else-if="solo") {{ tournamentStarted ? "Players" : "Registered players" }}
+      h2(v-else) {{ tournamentStarted ? "Teams" : "Registered teams" }}
+      .teams(v-if="teams.length" :class="{ solo }")
+        .team(
+          v-for="team in teams"
+          :class="{ myTeam: team === myTeam, ['rank-' + team.rank]: tournament.status === Status.Finished }"
+        )
+          .rank(v-if="tournament.status === Status.Finished") {{ team.rank }}
 
-template(v-for="(teams, index) in teamsGroups")
-  h2(v-if="tournament.status === Status.Finished") Results
-  h2(v-else-if="teamsGroups.length > 1 && index === 0") Incomplete teams
-  h2(v-else-if="tournament.teamSizeMax > 1") {{ tournamentStarted ? "Teams" : "Registered teams" }}
-  h2(v-else) {{ tournamentStarted ? "Players" : "Registered players" }}
-  table.teams(v-if="teams.length")
-    tr(v-for="team in teams")
-      td.rank(v-if="tournament.status === Status.Finished") {{ team.rank }}
-      td.team
+          .info
+            template(v-if="!solo || !team.members.length")
+              header
+                .name {{ team.name }}
+                .actions(v-if="!tournamentStarted")
+                  button(v-if="team === myTeam" @click="removeMember(team, myself)" title="Leave team")
+                    location-exit
+                  button(v-if="team === myTeam || isOrganizer" @click="renameTeam(team)" title="Rename team")
+                    pencil
+                  button(v-if="isOrganizer && !isTeamFull(team)" @click="addMemberPrompt(team)" title="Add member")
+                    account-plus
+                  button.icon.red(v-if="isOrganizer" @click="deleteTeam(team)" title="Delete team")
+                    delete
+              ul.members(v-if="team.members.length")
+                li.member(v-for="member in team.members")
+                  .user.icon-text(:class='{ myself: member.userId === user.id }')
+                    account
+                    user-link(:user="member.user")
+                  .actions
+                    button.icon.red(
+                      v-if="isOrganizer && !tournamentStarted"
+                      @click="removeMember(team, member)"
+                      title="Remove from team"
+                    )
+                      account-remove
+              p.noMembers(v-else) There are no members in this team
+              footer(v-if="canJoin(team)")
+                .membersCount
+                  account
+                  span {{ team.members.length }} / {{ tournament.teamSizeMax }}
+                button.button(@click="addMember(team, myself)")
+                  location-enter
+                  span Join
 
-        template(v-if="tournament.teamSizeMax > 1 || !team.members.length")
-          span.name {{ team.name }}
-          template(v-if="!tournamentStarted")
-            button(v-if="canRegister && !isRegistered && !isTeamFull(team)" @click="addMember(team, myself)") Join
-            button(v-if="team === myTeam || isOrganizer" @click="renameTeam(team)") Rename
-            button(v-if="team === myTeam" @click="removeMember(team, myself)") Leave
-            button(v-if="isOrganizer && !isTeamFull(team)" @click="addMemberPrompt(team)") Add member
-            button(v-if="isOrganizer" @click="deleteTeam(team)") Delete
-          ul.members(v-if="team.members.length")
-            li.member(v-for="member in team.members")
-              user-link(:user="member.user")
-              button(v-if="isOrganizer && !tournamentStarted" @click="removeMember(team, member)") Remove
-          p(v-else) There are no members in this team
+            .member(v-else)
+              .user.icon-text(:class='{ myself: team.members[0].userId === user.id }')
+                account
+                user-link(:user="team.members[0].user")
+              .actions
+                button.icon.red(v-if="isOrganizer && !tournamentStarted" @click="deleteTeam(team)" title="Unregister")
+                  account-remove
 
-        template(v-else)
-          user-link(:user="team.members[0].user")
-          button(v-if="isOrganizer && !tournamentStarted" @click="deleteTeam(team)") Remove
+          .result(v-if="tournament.status === Status.Finished") {{ team.result }} pts
 
-      td.result(v-if="tournament.status === Status.Finished") {{ team.result }} pts
-
-  p(v-else) Nobody has registered to this tournament yet.
+      p(v-else) Nobody has registered to this tournament yet.
 </template>
 
 <style scoped lang="sass">
-.teams td
-  vertical-align: text-top
+@import "../assets/styles/main"
+@import "../assets/styles/tournament"
 
-.team
-  .name
+.container
+  display: flex
+  flex-direction: column
+  margin-bottom: 20px
+
+.tournament-header
+  flex-shrink: 0
+  display: flex
+  padding: 16px 20px
+  align-items: center
+  position: relative
+  z-index: 1
+  overflow: hidden
+  box-shadow: 0 15px 20px rgba(0, 0, 0, 0.15)
+
+  .bg
+    background-size: cover
+    background-position: center
+    position: absolute
+    left: 0
+    top: 0
+    width: 100%
+    height: 100%
+    opacity: 40%
+    z-index: -2
+
+  .tournament-title
+    flex: 1
+    text-shadow: 0 0 10px rgba(0, 0, 0, 0.5)
+
+    h1
+      font-weight: normal
+      margin: 0
+
+    .status
+      color: #bbd
+      margin: 8px 0 0
+
+  .teamCount, .register, .unregister
+    display: flex
+    flex-direction: column
+    align-items: center
+    padding: 10px
+    border-radius: 5px
+    margin-left: 16px
+
+    svg
+      margin-bottom: 8px
+      font-size: 2em
+
+  .teamCount
+    background-color: rgba(0, 0, 0, 0.3)
+    width: 80px
     font-weight: bold
 
+    .low
+      color: #ee6391
+
+  .register, .unregister
+    width: 100px
+
+  .unregister
+    @extend .red
+
+  .admin
+    background-color: rgba(0, 0, 0, 0.3)
+    border-radius: 5px
+    align-self: stretch
+    display: flex
+    flex-direction: column
+    justify-content: space-between
+    align-items: center
+    padding: 8px
+
+    span
+      font-size: 0.8em
+
+    .button:not(:last-child)
+      margin-right: 8px
+
+.started
+  .tournament-header
+    background: $running-bg
+
+  .tournament-title .status
+    color: palegreen
+
+.content
+  display: flex
+
+.rules
+  max-width: 700px
+  padding: 15px
+  margin: 0 15px
+  flex: 1
+
+.teamsGroup
+  $width: 230px
+  flex: 0 0 $width
+  max-width: $width
+  padding-right: 20px
+
+  h2
+    @extend .skewed
+    text-align: center
+
+  &:not(:last-child) .team
+    background-color: rgba(0, 50, 0, 0.3)
+
+.finished .teamsGroup
+  $width: 300px
+  flex: 0 0 $width
+  max-width: $width
+
+.team
+  display: flex
+  align-items: baseline
+  background-color: rgba(0, 0, 0, 0.2)
+  padding: 10px
+  margin: 15px 0
+  border-radius: 5px
+
+  .rank
+    font-weight: bold
+    margin: 2px 10px 2px 0
+    width: 24px
+    height: 24px
+    flex-shrink: 0
+    color: lightblue
+    display: flex
+    align-items: center
+    justify-content: center
+
+  .result
+    margin-left: 6px
+    white-space: nowrap
+    flex-grow: 1
+    text-align: right
+    font-size: 0.9em
+
+  .info
+    width: 100%
+    overflow: hidden
+
+    header
+      display: flex
+      justify-content: space-between
+      align-items: center
+      height: 26px
+
+  .name
+    font-weight: bold
+    overflow: hidden
+    white-space: nowrap
+    text-overflow: ellipsis
+
+  .actions
+    white-space: nowrap
+
+    button
+      @extend .icon
+      opacity: 0
+      transition: opacity 0.1s linear
+
+      &:not(:last-child)
+        margin-right: 5px
+
+  &:hover .actions button
+    opacity: inherit
+
   .members
+    padding-left: 6px
+    margin: 6px 0
+
+  .member
+    display: flex
+    justify-content: space-between
+    align-items: center
+    height: 30px
+
+    .user
+      max-width: 100%
+
+      a
+        overflow: hidden
+        white-space: nowrap
+        text-overflow: ellipsis
+
+    .user:not(.myself) svg
+      opacity: 50%
+
+  footer
+    display: flex
+    justify-content: space-between
+    align-items: baseline
+
+  .membersCount
+    @extend .icon-text
+    color: rgba(255, 255, 255, 0.8)
     padding-left: 15px
+
+  .noMembers
+    font-size: 0.8em
+    font-style: italic
+
+.teams.solo
+  background-color: rgba(0, 0, 0, 0.2)
+  border-radius: 5px
+  padding: 10px
+
+  .team
+    background-color: initial
+    align-items: center
+    margin: 0
+    padding: 0
+
+:not(.solo) > .myTeam
+  outline: 2px solid rgba(200, 220, 255, 0.5)
+  background-color: rgba(0, 0, 50, 0.3)
+
+.finished .team .info
+  width: unset
+
+.rank-1 .rank, .rank-2 .rank, .rank-3 .rank
+  font-size: 1.1em
+  border-radius: 100%
+
+.rank-1 .rank
+  background-color: #ffe989
+  color: #bd7b32
+
+.rank-2 .rank
+  background-color: #b4c1c7
+  color: #425777
+
+.rank-3 .rank
+  background-color: #b48b7e
+  color: #721b0d
 </style>
